fix: change directory
